Don't use leases for silenced entries (#4034)
This commit changes how silences work in Sensu.

Before, silences were governed by etcd leases. The idea was that a
silenced entry would be created and leased to expire, if it had the
Expire field set.

This was fine in theory, and often in practice, but leases have a subtle
characteristic - they are reset when etcd holds an election, if they are
less than 5 minutes long.

This can cause havoc when Sensu is restarted. For a feature like
silences, it is not reliable enough.

Now, silences store their expiry information directly in their own data
structure. When silences are listed, expired results are simply
discarded from the result set, and asynchronously garbage collected.
This should make silences perfectly reliable, as they will no longer
rely on any state in etcd or Sensu, but rather be self-contained in
their own data structure.

The core/v2.Silenced datatype has a new field, called ExpireAt, which is
the time in seconds since the Unix epoch. When this field is set, Sensu
will use it to determine if the entry has expired. The field will be
automatically set by the store when a new silence is created, and will
therefore be API-compatible with the previous implementation.

If the ExpireAt field is not set on a silenced entry read from the
database, then Sensu will assume that the entry pre-dates the new system,
and fall back to requesting its lease.

Signed-off-by: Eric Chlebek <eric@sensu.io>