Refactor the round-robin ring and consumers
The commit refactors the round-robin ring so that it can be abstracted
away by its consumers.

The existing interfaces for the ring remain, but new facilities have
been added (ringv2.Interface, ringv2.RingPool).

The new RingPool allows patching the internal ring implementation from
outside, so that software like sensu-enterprise-go can replace the
implementation for another one.

This also should allow us to mock the ring in tests, however, at this
time, this has not been done.

Signed-off-by: Eric Chlebek <eric@sensu.io>