Unconditionally cancel ring watchers
This commit causes ring watchers to cancel unconditionally when rings
are updated.

Previously, the scheduler tried to be optimal about this, avoiding
cancelling any ring watchers if the perceived schedule had not changed.
However, now that we have pluggable schedulers, this logic is faulty and
must be removed.

This should not change the behaviour of round-robin checks for OSS and
may even improve the situation for backends that are experiencing issues
with round-robin scheduling.

Signed-off-by: Eric Chlebek <eric@sensu.io>