Remove managed_by label creation with edit command
During a previous refactor, parts of the logic for the create, edit and
delete sub-commands were merged together. This resulted in the side
effect of having the edit and delete sub-commands automatically add a
"managed_by" label with a "sensuctl" value to resources they touched.
This is not what we want; only sensuctl create should be automatically
adding this label.

This commit removes the label creation from the code shared by the 3
sub-commands, creates a new Processor that labels resources before
submitting them to the API and uses it as part of the definition of the
create sub-command only.

Signed-off-by: Cyril Cressent <cyril@sensu.io>