Add support for core/v3.Resource to backend/api
This commit adds support for core/v3.Resource types in the backend API
GenericClient. The support is shimmed through the
core/v3.V2ResourceProxy type.

Users are expected to set up the GenericClient with a store/v2.Interface
and a Kind that is a V2ResourceProxy, if they want to work with core/v3
resources.

Signed-off-by: Eric Chlebek <eric@sensu.io>