Fix CreatedBy not being carried over when converting v3 to v2 entities (#4133)
With this change, when converting a (v3.EntityConfig, v3.EntityState)
into a v2.Entity, we carry over the CreatedBy metadata from the config,
just like we do for labels and annotations.

After careful consideration, a test that asserts that, after a v3 to v2
conversion, the state's metadata ought to be the same as the resulting
v2 entity's has been removed because it is not correct: metadata
information is copied from the config part. This change exposed this
issue.