Only calculate build info once
This commit fixes a bug where build information is calculated for each
keepalive generated. This generates some additional work for the agent
which should be marginal in all but the most extreme cases.

The bug manifests itself more clearly when using the loadit tool for
load testing, consuming a large fraction of the CPU when 10,000 agents
are loaded.

Signed-off-by: Eric Chlebek <eric@sensu.io>