Make SIGHUP a no-op (#4218)
This commit removes the functionality that causes sensu-go's internal
daemons to restart on SIGHUP. Instead, SIGHUP is now a no-op.

In the enterprise product, SIGHUP is used to rotate the special event
log file, so the restarting behaviour was undesirable.

We don't anticipate anyone should ever want to force an internal daemon
restart, users should instead restart the whole sensu-backend process if
that's what they want to do. So, support for doing this is now removed.

Signed-off-by: Eric Chlebek <eric@sensu.io>