Add labels to canned events for load testing
This commit adds a static set of labels to each event produced in load
testing scenarios. The intent is to make the events a little more
realistic in terms of their payload. Labels are indexed by postgres in
the commercial distribution, and may have an impact on performance that
is currently unmeasured.

Signed-off-by: Eric Chlebek <eric@sensu.io>