Remove EventStoreProxy, add StoreProxy
This commit removes EventStoreProxy and replaces it with StoreProxy.
StoreProxy is like EventStoreProxy, but allows replacement of the entire
Store.

Unlike EventStoreProxy, StoreProxy uses a sync.RWMutex for
synchronization, instead of a pointer swap. This should be fast enough
for the store, while removing the use of package unsafe.

Projects that relied on EventStoreProxy should adapt their usage of it
to StoreProxy, which can be done relatively easily.

Signed-off-by: Eric Chlebek <eric@sensu.io>