Add storev2.WrapResource, and wrap.PreferPostgres
This commit adds a replaceable function, storev2.WrapResource, that
clients can call instead of wrap.Resource. It allows patching the
wrapper function from sensu-enterprise-go, so that it can optionally
prefer to use Postgres.

Preference for using postgres is done on a case by case basis in
sensu-go, currently only for entity state. To facilitate this, a new
wrapper option has been added to the wrap package, called
PreferPostgres.

Signed-off-by: Eric Chlebek <eric@sensu.io>