Don't emit fatal ring errors from keepalived
This commit changes the behaviour of keepalived such that roundrobin
ring errors are now logged but don't result in an internal restart being
triggered.

Under extreme load, etcd rings can be highly unstable, but it's usually
preferable to continue on when there are lease faults but K/V operations
are still generally working.

Signed-off-by: Eric Chlebek <eric@sensu.io>