Add support for javascript mutators (#4361)
This commit adds a Type field to core/v2.Mutator which can be used to
specify if the mutator is a pipe mutator, or a javascript mutator. If
the Type field is unset, then the default is pipe.

Javascript mutators can operate on a pre-defined object called "event".
Any mutations made to this object will be propagated through the
pipeline. Alternatively, a wholly new value can be returned, as long
as it of type string.

Runtime assets are fully supported, so long as they are valid javascript
assets.

Timeouts are supported, but tight loops may block the runtime system. Be
mindful of the mutator code you write.

Environment variables are supported and are made available as global
variables in the execution environment.

Secrets are not supported.

Minimal example:

api_version: core/v2
type: mutator
metadata:
  namespace: default
  name: doge
spec:
  command: 'event.check.labels["doge"] = "woof";'
  type: javascript

Signed-off-by: Eric Chlebek <eric@sensu.io>