Add platform metrics logging and event logging
This commit adds platform metrics logging to sensu-go. The feature
defaults to being enabled, but can be disabled via flag, config file, or
environment variable. See documentation for details.

This commit also adds event logging, previously an enterprise-only
feature, to OSS sensu. The flag names and behaviours for this feature
are the same as in the enterprise product.

Both of these logs can be rotated by sending SIGHUP to sensu-backend
after mv-ing the log files.

Signed-off-by: Eric Chlebek <eric@sensu.io>