Partition silenced entry transactions
The number of silenced entries in a given transaction could exceed the
default number of etcd transaction operations, which is 128. This commit
changes the silenced entry store so that it makes at most 64 transaction
operations, and handles silenced expiration by creating multiple
transactions that are each no longer than 64 items.

Signed-off-by: Eric Chlebek <eric@sensu.io>