Expose env vars to javascript mutators
This commit allows exposing environment variables to javascript mutators
by specifying them in the mutator env_vars section, but without an "="
sign.

For example, in this configuration,

{"env_vars": ["FOO", "BAR=BAZ"]}

The mutator would have a global variable "BAR" set to the value "BAZ",
and a global variable "FOO" set to the value of FOO in the
sensu-backend's environment.

Note: this is inconsistent with a regular mutator, which has access to
environment variables by virtue of being a child process of sensu-backend.

Signed-off-by: Eric Chlebek <eric@sensu.io>