Thread contexts properly throughout sensu-backend
Due to the legacy daemon.Daemon interface, contexts were never properly
threaded throughout sensu-backend's daemon components. As a result, many
components use contexts internally and rely on a somewhat twisted
shutdown method.

This tends to be buggy. Because of this design we've had several
instances where sensu-backend won't shut down when asked to do so. The
solution is to change the daemon.Daemon interface so that Start accepts
a context.Context. The behaviour of Stop() changes as well, with Stop()
blocking until after the context provided to Start() is cancelled.

Signed-off-by: Eric Chlebek <eric@sensu.io>