Do less work in agent session shutdown
The agent session manager tries to remove the agent from each ring it is
involved in. While this is worthwhile, it also creates a lot of load all
at once on a sensu-backend that is in the process of shutting down. This
commit changes our posture towards this.

Instead of trying to remove the agent from each ring it is involved in,
only do this work if the agent is configured to deregister on session
shutdown. We can reasonably do this when we expect the agent to re-join
the cluster.

This commit also stops processing ring removals in parallel, which seems
to place too much load on etcd, resulting in rate limiter errors.

Signed-off-by: Eric Chlebek <eric@sensu.io>