Fix several websocket handling bugs
This commit changes the way websocket connections are handled, by both
the backend and the agent.

It replaces one sync.RWMutex with two sync.Mutex values, one for reads,
and one for writes. The gorilla websocket docs state that reads and
writes can be concurrent, so it's not correct to use a single mutex for
both resources.

The commit also avoids setting the internal "closed" flag to true when a
connection error is encountered. Instead, it only sets the flag to true
when a ClosedError is found on receive.

The commit also makes sure the write mutex is locked when sending a
close message.

The commit also removes sync.Pool use for transport messages, which was
a speculative optimization we applied. However, it seems unlikely that
this optimization is really necessary, so I have removed it in service
of making the code clearer.

Signed-off-by: Eric Chlebek <eric@sensu.io>