Try setting grpc.FailFast on the etcd client
The default for etcd clients is to do this, but it's not clear if that
will happen when the DialOptions are explicitly configured. So to make
sure we're working the way the etcd server expects, set
grpc.FailFast(false) in the DialOptions.

grpc.FailFast is deprecated by a new call option, grpc.WaitForReady,
which has the inverse meaning, but I think etcd are still using the
deprecated option, so stubbornly, I will too.

Signed-off-by: Eric Chlebek <eric@sensu.io>