[GraphQL] Use a sliding chunk size (#4599)
Switches to using a chunk size of between 250 and 500 records. In cases where the page size is small (eg. 25 items) this will lead to a bit more memory usage but it greatly reduce the amount to round trips. As I/O is likely a bigger bottleneck than CPU or memory, this should greatly reduce resolver execution time.

Signed-off-by: James Phillips <jamesdphillips@gmail.com>