Fix a panic in socket handlers (#4678)
This commit fixes a case in socket handlers that can cause sensu-backend
to crash with a nil pointer dereference. The condition occurs when
handlers can connect to a socket, but then get an error while writing.

Added a unit test that fails if applied prior to this commit.

Unfortunately adding this failing unit test was not overly simple. I had
to fix a bug where timeout was not respected by Write(). I also
refactored the unit tests because they all bind to the same port. They
now bind to pseudo-random ports.

Signed-off-by: Eric Chlebek <eric@sensu.io>