Remove json-iterator from the project
Recently we found that json-iterator can panic due to incorrect use of
package unsafe. Since json performance is less of a concern than it used
to be, and since encoding/json has progressed in its performance since,
then, I believe it's wise to remove it from the project altogether.

I've also created new tags for core/v2, corev/3, and types modules, and
have updated the go.mods in the repository to reflect that.

Signed-off-by: Eric Chlebek <eric@sensu.io>