Remove support for embedded etcd (#4704)
* Remove embedded etcd clustering

This is a substantial breaking commit. It removes several etcd
configuration options for configuring embedded etcd clustering. At one
time, this was the only way that sensu-go worked.

We've learned that embedded etcd introduces a lot of operational
fragility, and are now mandating that etcd be an external dependency.

To improve hackability and developer velocity, there is now a
zero-dependency developer mode for sensu-go. The developer mode makes
use of a single-node etcd instance, but we reserve the right to change
it to something else in the future of the 7 series releases.

This is the first in a series of commits that will add full support for
PostgreSQL to sensu-go. We have chosen to de-emphasize etcd in favour of
PostgresQL, as we have found that it is far easier for users to deploy
PostgresQL on cloud infrastructure (and possibly also on-prem
infrastructure).

In addition to these breaking changes, some routine gardening has been
done. The concept of integration build tags has been removed.
Integration test build tags are troublesome for static analysis tools
and editors that compile on save. We introduced them at a time that Go
didn't have test caching, and now it is time to de-introduce them. The
whole senus-go test suite can run with the race detector enabled in just
a few minutes.

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Simplify build.sh and build.ps1

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Fix mangled .circleci.yml

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Update CHANGELOG

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Fix several data races in schedulerd

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Remove unused code

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Alias dev mode etcd server log level to backend log level

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Remove appveyor integration test suite

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Only test agent and sensuctl on Windows (and their deps)

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Mitigate a data race in agentd sessions

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Add cluster version info for dev mode

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Don't have a timeout for seeding tests

Signed-off-by: Eric Chlebek <eric@sensu.io>

* Log that cron is stopping before stopping it

Signed-off-by: Eric Chlebek <eric@sensu.io>