Add postgres event store and selector library
This commit adds a significant volume of code from the enterprise Sensu
product. In particular, it adds the postgres state store, and the
selector library that enables selector expressions.

The postgres state store contains implementations of:
- store.EntityStore
- store.EventStore
- ringv2.Interface

The current form of the event store is historical and unlikely to
survive 7.0's planned work, but in the interest of incremental change
it's copied over faithfully.

These changes are not enough to get a working postgres state store in
sensu-go. More work is needed to integrate these components with the
wider system. The code that enables this in the enterprise product is
coupled tightly with other things, so some porting work needs to happen
before any of this code can be employed.

Signed-off-by: Eric Chlebek <eric@sensu.io>