Add Operator Presence Control system
This commit adds a new persistent state manager to sensu-go called
Operator Presence Control. Operator Presence Control or OPC is an
accounting system for operator presence. An operator can be any entity,
system or condition that has a notion of presence. For instance, agents
can use the OPC system to implement keepalives, and the check scheduler
can use the OPC system to implement check TTL.

When an operated is checked in, a row in the OPC SQL table is inserted
or updated with the operator's details, the current time, and the
duration that can elapse without further check-ins before the operator
is considered to be absent.

Operators can be assigned controllers. Controllers are also operators,
and are useful for sharding monitoring work. For instance, agent
operators can be assigned backend operators as controllers. Each backend
can then concern itself only with the agent operators that are assigned
to it.

Controllers can become absent too. When that happens, the monitoring
system will reassign operators missing a controller to the other
controllers of the same type in the system. The operators will be
equally balanced among all of the controllers.

Operators can also assign themselves a different controller by checking
in with a different controller in their check-in request.

The OPC system differs from the existing notification systems in
sensu-go. Notably, the system depends on postgresql instead of etcd, and
uses polling instead of asynchronous notification. This should permit a
much broader use of these types of notifications, since all notification
events are handled in a batch within the same goroutine. We expect
keepalive storms and check TTL storms to be much less damaging to system
stability with this design.

Signed-off-by: Eric Chlebek <eric@sensu.io>