Agent ingests graphite plaintext metric tags (#4897)
Graphite has support for storing data in tags to identify series of
metrics. This change introduces support for this feature in the graphite
text protocol. Presently the agent ingests
`my.series;tag1=value1 1 999999999` as a metric point with name
"my.series;tag1=value1". Per the graphite spec this will now be ingested
as a metric point with name "my.series" and a tag for ("tag1": "value1")

Signed-off-by: Christian Kruse <ctkruse99@gmail.com>

reframe changelog as a feature add

Signed-off-by: Christian Kruse <ctkruse99@gmail.com>

Signed-off-by: Christian Kruse <ctkruse99@gmail.com>