Make check TTL work again
This commit makes check TTL work with the new OPC system. Each check
that specifies the TTL field will have an entry stored in the OPC table
for each agent entity that executes the check.

A new OPC monitoring option has been added to the
MonitorOperatorsRequest struct, which allows selecting all of the
grandchildren of the requested controller operator instead of the
children.

Signed-off-by: Eric Chlebek <eric@sensu.io>