Use wrapped resources in all APIs
The sensu-go APIs have been operating on raw/unwrapped json encoded
resource definitions. This is a common source of friction as most users
are primarily familiar with our wrapped resource format that sensuctl
presents to users as the interface.

This commit changes all sensu-go APIs dealing with resources to operate
on the wrapped resource format instead.

Signed-off-by: Christian Kruse <ctkruse99@gmail.com>